[preanalysis] Adding a preanalysis to resolve local closures
Summary:
This will allow all the analyses to be able to call closures without any special treatment: we transform the call to variables that point to closures into normal function calls. We treat only ObjC blocks at the moment, with C++ lambdas to be done as a next step.

We aimed to achieve certain results in Pulse (see tests: avoid memory leaks and NPEs FPs) while also keeping the biabduction analysis working as before.

We also checked that for the examples analyzed Pulse behaves like the correct semantics of ObjC programs with blocks.

Reviewed By: jvillard

Differential Revision: D22547333

fbshipit-source-id: efe56ed51