[absint][pulse] Remove NeverJoin
Summary: There used to be `JoinAfter n` mode where we would try to join `n` states instead of always making disjunctions. It got deleted in D14258485 and Pulse's underlying (pre-disjuncts) domain doesn't even have a join operation.  `NeverJoin` mode is not useful in Pulse anymore: pulse will diverge or OOM if we don't limit the number of disjuncts. It is also not used by any other analyzer.  Let's remove it.

Reviewed By: jvillard

Differential Revision: D22817425

fbshipit-source-id: 1e658f11d