[nullsafe] Add a test for @SuppressFieldNotInitialized
Summary:
There are two ways to suppress it.
1. Field level suppression annotation (was already tested). This will
apply to all constructors.
2. Constructor level annotation (this is what this test does). Sometimes
there are "fake" constructors that are not intended to be called in
prod, they might leave some fields not initialized.

Note that there are two ways to add a suppression; one has a known
problem that is documented here.

Reviewed By: artempyanykh

Differential Revision: D22864869

fbshipit-source-id: f95aaa26a