[java][procname] params are Typ.t
Summary:
Use `Typ.t` for parameter types in procnames, instead of `JavaSplitName.t`.  This allows more precise and usable types stored in procnames, as well as not using the string-based representation of `JavaSplitName` which meant parsing strings whenever one needed a `Typ.t`.

This also makes `JavaSplitName` dead code, so it is removed.

Reviewed By: skcho

Differential Revision: D20500423

fbshipit-source-id: a72728e3f