[frontend] Remove duplicate variable declaration
Summary:
We get duplicated variable declaration instruction for primitive type variable initialized using list initializer, e.g.
```
int* p{nullptr};
```
This happens because we add variable declaration instruction when we translate both `DeclStmt` and `InitListExpr`. To fix this, we do not add the duplicated variable declaration when we translate `InitListExpr`.

Reviewed By: jvillard

Differential Revision: D22844726

fbshipit-source-id: 422806924