[frontend] Record attribute unused in local variable data
Summary: To avoid dead store false positives we skip initialization of a variable that has an `unused` attribute. However, this causes uninitialized value false positives when the variable is later used in macros. To fix this, instead of skipping initialization we record the information about `unused` attribute in local variable data that we can later use for filtering out dead store issues.

Reviewed By: jvillard

Differential Revision: D22868050

fbshipit-source-id: 4a2d0e680