[nullsafe] @Initializer annotation is inherited from parent signatures
Summary:
We already take into account inheritance if the method inherits a known
modelled initializer method (e.g. Activity.onCreate()).

But if the method is explicitly marked as Initializer, we require its
overrides to be also marked.

This diffs fixes the behavior and makes it consistent, now this is
enough to annotate only parent class.

Reviewed By: jvillard

Differential Revision: D23135177

fbshipit-source-id: a21ff4a0e