[pulse] preserve physical equality on var subst in LinArith
Summary:
This is used for variable substitution and will often be a no-op when
normalising terms over and over again (after the first normalisation,
the expression should stay the same). The equivalent function for terms
was already being careful about not re-allocating identical terms so
extend that care to linear expression.

Reviewed By: skcho

Differential Revision: D23241601

fbshipit-source-id: b365eb87a