[closures preanalysis] [1/n] Adding first step of preanalysis to specialize methods with concrete closures they are called with
Summary:
This preanalysis in general aims to create specialized clones of methods that have blocks as arguments and that are called with concrete closures, and then call these clone methods instead of the original ones.

One complication is what happens with the captured variables in the closure. What we do is we add them to the formals of the cloned method and passed them through to the concrete blocks.

We do this transformation in two steps:
1. Go through all the callers of methods with blocks as parameters, and create the clone methods. In this preanalysis we only create the attributes for the new method, not the code. We also update the call instructions in the callers to represent a call to the cloned method with updated arguments: we don't need to pass closures arguments anymore, we instead pass the captured variables as new arguments.
2. We add the corresponding code to the newly created clones: this means swapping the call to the block variable with a call to the corresponding block.  Moreover, we add some of the new formals (that correspond to the captured variables) to the arguments of the call.

This diff implements step 1 of the analysis.  The next diff D23216021 implements step 2.

Reviewed By: ngorogiannis

Differential Revision: D23109204

fbshipit-source-id: 91a5eb16b