[biabduction][models] move to sqlite
Summary:
Store model summaries in the `model_specs` database table instead of in spec files.

This table is populated when a new database is created by loading a dump of the `specs` table in the models database.  This avoids the perf and reliability implications of ATTACHing the same, non-read-only models-DB by many processes.

- `BiabductionModels` is moved into `IR` so that `JsonReports` can access it.
- The binary `sqlite3` is now required on the host compiling infer.

Reviewed By: skcho

Differential Revision: D23191601

fbshipit-source-id: 1532481ee