[sledge] Add: SMT-LIB frontend to SLEdge's internal first-order theory solver
Summary:
The SLEdge internal first-order theory solver targets the "word
problem", where a query has the form `C ⊢ L` where `C` is a
conjunction of literals and `L` is a single literal. This can be
abused to implement a naive SMT solver using an expansion to
disjunctive-normal form and checking `C ⊢ false` for each branch `C`
of the DNF. This is not useful as an SMT solver, but can be used for
testing.

Reviewed By: ngorogiannis

Differential Revision: D23459524

fbshipit-source-id: 5483e5a84