[sledge] Improve: Replace naive implementation of Fol.equal_or_opposite
Summary:
`Fol.equal_or_opposite p q` naively constructs the negation of `q` in
most cases, only to test if it is equal to `p`. This is an inefficient
method of computing if one formula is the negation of another. This
diff implements this directly. The drawback is some duplication of the
negation logic from `_Not`.

Reviewed By: ngorogiannis

Differential Revision: D23487500

fbshipit-source-id: 100f95edc