[attributes] request attributes only when procdesc isn't required
Summary: Most of the time, when the procdesc of a callee is requested, all that is really required is the procedure attributes. However, requesting the procdesc may return `None` when the procedure is undefined (in Java, and soon for Clang too).  So, change all callsites to using attributes instead, where possible.

Reviewed By: jvillard

Differential Revision: D23539422

fbshipit-source-id: 3b1a52d48