[cost] Revise CostDomain.plus functionality: zero+unreachable
Summary:
We defined cost's plus as "zero+unreachable = unreachable" for the operation cost.  The meaning of
the zero cost is that no statment is analyzed yet, and the unreachable(bottom) cost means a program
point is analyzed as unreachable.  We thought "zero+unreachable" happens in very specific cases we
need to check, or due to an analyzer bug.  For debugging purpose, we defined it to return more
specific unreachable cost.

However, in allocation/autoreleasepool costs, the zero cost is not very special value. If there is
no alloc/autorelease calls, they can have the zero cost. "zero+unreachable = unreachable" doesn't
make sense there.

This diff changes the plus as "zero+unreachable = zero" for the non-operation costs.

Reviewed By: ezgicicek

Differential Revision: D23578869

fbshipit-source-id: 5392eca1c