[scheduler][restart] use proc_uids instead of serialised procnames as targets
Summary:
Eliminate the need to serialise procnames when sending work from the restart scheduler to the workers, by sending the proc_uid instead.  This is (much) shorter than the byte representation of the proc_name and it's the primary DB key of the procedures table, so it can be used by the worker to obtain the full procname.

Also, reduce GC churn by using folds in the scheduler startup instead of copying lists over and over.

Reviewed By: jberdine

Differential Revision: D23566131

fbshipit-source-id: 1472aa990