[cost] Include the previous cost's trace in differential traces
Summary:
For complexity issues from O(m) to O(n), we only include the trace of the current complexity O(n). However, this makes it difficult to understand what the original complexity O(m) was. Especially in fixed issues where n=1, we only get a constant cost with no trace attached, so it is difficult to see how the symbol m disappeared.

This diff includes the traces for the previous cost in the cost issues.

Reviewed By: skcho

Differential Revision: D23680360

fbshipit-source-id: 3f2b21b20