[cost] add differential cost test for objc
Summary:
When inspecting silent introduced issues, we found that an introduced issue is about a complexity increase of a block that is only created in the current diff. Based on the trace view, we find out that this is caused by infer mistakingly consider another block that exists in the previous diff as the same block that is newly created in the current diff.

This diff adds a test case that reproduces this case, and this will be fixed in the next diff.

facebook
Trace view: https://www.internalfb.com/intern/traceview/?id=109896337

Reviewed By: ezgicicek

Differential Revision: D23681550

fbshipit-source-id: 78341268b