[cost] fix objc block name comparison
Summary:
Following the previous diff, the reason we mistakingly introduced new block as complexity increases is because when we are computing the hash of block, we return the string `block` to compute the hash for all block.

This diff fixes this case by returning the non-verbose name of the block when returning hashable name.

Reviewed By: ezgicicek

Differential Revision: D23729710

fbshipit-source-id: a345d3045