[SIL] Change WithBlockParameters to have Block.t
Summary:
`WithBlockParameters` is generated by a pre-analysis to express concrete block parameters.  However,
before this diff, the block parameters only have names, which is insufficient to find their
summaries.  This diff change `WithBlockParameters` to have `Block.t` that includes the parameters of
the block, so we can find blocks' summaries.

Reviewed By: ezgicicek

Differential Revision: D23785148

fbshipit-source-id: 9034f4f8d