[closure] Keep closure parameter in specialized procedure
Summary:
This diff keeps closure parameters in closure-specializated procedures.

What the closure-specialization is doing is a propagation of concrete closures.  For example, it
translates:

```
foo(block b) {
  b();
}

goo() {
  foo(^{...});
}
```

to

```
foo_new() {
  (^{...})();
}

goo() {
  foo_new();
}
```

However, if `foo` addresses `b` as a normal value like

```
foo(block b) {
  block c = b;
}
```

this is translated to

```
foo_new() {
  block c = b;
}
```

Note that the closure parameter of `foo` is removed, thus `b` becomes a free variable.  Not good.
To avoid the situation, this diff keeps the closure parameters intact.

Reviewed By: da319

Differential Revision: D23905580

fbshipit-source-id: 014989fbf