[pulse] report errors only when the PRUNE nodes along the path are true
Summary:
Take another page from the Incorrectness Logic book and refrain from reporting issues on paths unless we know for sure that this path will be taken.

Previously, we would report on paths that are merely *not impossible*. This goes very far in the other direction, so it's possible we'll want to go back to some sort of middle ground. Or maybe not. See the changes in the tests to get a sense of what we're missing.

Reviewed By: ezgicicek

Differential Revision: D24014719

fbshipit-source-id: d451faf02