[cost] Extend the polynomial domain for closure field
Summary:
This diff extended the polynomial domain to include symbols for closure calls.

When the closure symbol is added to the polynomial? Unknown closure is called inside a function
like,

```
foo() {
  self->closure_field();
}
```

Thus, the cost of `foo` becomes `|self->flosure_field|`, rather than unknown. (Note that this
semantics is added only for autoreleasepool size at the moment.)

When the symbol is instantiated? `foo` is called with correct closure contexts.

```
goo() {
  self->closure_field = ^(){ ... };
  foo();
}
```

The summary of `goo` will have instantiated summary of the closure.

Reviewed By: ezgicicek

Differential Revision: D23992590

fbshipit-source-id: d1d228403